Add 'distribution' metrics
This is an extensions to the statsd client that is specific to Datadog and enables the gathering of 'Distribution' metrics.  More information about distribution metrics can be found here:

https://docs.datadoghq.com/metrics/distributions/

Since this is a non-standard datadog extension to statsd I looked at their 'official' implementation to determine the byte type used for distributions.  Found it here:
   https://github.com/DataDog/datadog-go/blob/master/statsd/format.go#L12